added decimal prices, Federal Tax and On demand Provincial Tax handling

+++ pos/pos/IDisplay.cs
@@ -3,6 +3,6 @@ namespace pos
 {
     public interface IDisplay
     {
-        void PrintPrice(int iVal);
+        void PrintPrice(decimal iVal);
     }
 }
